Rename to ChapterEntryBlock
