+++ data/test_chapters.json
@@ -32,7 +32,7 @@
           "slug": "test"
         }
       ],
-      "content": "# Test\nThis is the start of a sentence.",
+      "content": "# Getting started as a Freshman can be challenging and overwhelming. To help you find your way, we have curated a list of articles to help you get started on your academic journey.",
       "lastUpdated": "2023-02-23T20:40"
     }
   ]
